OO-501: separated models
