+++ src/main/scala/ch/openolitor/stammdaten/models/LieferungenModels.scala
@@ -40,19 +40,6 @@ object LieferungStatus {
   }
 }
 
-sealed trait KorbStatus
-
-case object WirdGeliefert extends KorbStatus
-case object Geliefert extends KorbStatus
-case object FaelltAusAbwesend extends KorbStatus
-case object FaelltAusSaldoZuTief extends KorbStatus
-
-object KorbStatus {
-  def apply(value: String): KorbStatus = {
-    Vector(WirdGeliefert, Geliefert, FaelltAusAbwesend, FaelltAusSaldoZuTief) find (_.toString == value) getOrElse (WirdGeliefert)
-  }
-}
-
 case class LieferplanungId(id: Long) extends BaseId
 
 case class Lieferplanung(
@@ -207,194 +194,4 @@ case class LieferpositionModify(
 case class LieferpositionenModify(
   preisTotal: Option[BigDecimal],
   lieferpositionen: List[LieferpositionModify]
-) extends JSONSerializable
-
-case class BestellungId(id: Long) extends BaseId
-
-case class Bestellung(
-  id: BestellungId,
-  produzentId: ProduzentId,
-  produzentKurzzeichen: String,
-  lieferplanungId: LieferplanungId,
-  status: LieferungStatus,
-  datum: DateTime,
-  datumAbrechnung: Option[DateTime],
-  preisTotal: BigDecimal,
-  steuerSatz: Option[BigDecimal],
-  steuer: BigDecimal,
-  totalSteuer: BigDecimal,
-  datumVersendet: Option[DateTime],
-  //modification flags
-  erstelldat: DateTime,
-  ersteller: PersonId,
-  modifidat: DateTime,
-  modifikator: PersonId
-) extends BaseEntity[BestellungId]
-
-case class BestellungDetail(
-  id: BestellungId,
-  produzentId: ProduzentId,
-  produzentKurzzeichen: String,
-  lieferplanungId: LieferplanungId,
-  status: LieferungStatus,
-  datum: DateTime,
-  datumAbrechnung: Option[DateTime],
-  preisTotal: BigDecimal,
-  steuerSatz: Option[BigDecimal],
-  steuer: BigDecimal,
-  totalSteuer: BigDecimal,
-  datumVersendet: Option[DateTime],
-  positionen: Seq[Bestellposition],
-  produzent: Produzent,
-  //modification flags
-  erstelldat: DateTime,
-  ersteller: PersonId,
-  modifidat: DateTime,
-  modifikator: PersonId
-) extends BaseEntity[BestellungId]
-
-case class BestellungAusgeliefert(
-  datum: DateTime,
-  ids: Seq[BestellungId]
-) extends JSONSerializable
-
-case class BestellungModify(
-  produzentId: ProduzentId,
-  produzentKurzzeichen: String,
-  lieferplanungId: LieferplanungId,
-  datum: DateTime,
-  datumAbrechnung: Option[DateTime],
-  preisTotal: BigDecimal
-) extends JSONSerializable
-
-@Deprecated
-case class BestellungenCreate(
-  lieferplanungId: LieferplanungId
-) extends JSONSerializable
-
-case class BestellungCreate(
-  produzentId: ProduzentId,
-  lieferplanungId: LieferplanungId,
-  datum: DateTime
-) extends JSONSerializable
-
-case class BestellpositionId(id: Long) extends BaseId
-
-case class Bestellposition(
-  id: BestellpositionId,
-  bestellungId: BestellungId,
-  produktId: Option[ProduktId],
-  produktBeschrieb: String,
-  preisEinheit: Option[BigDecimal],
-  einheit: Liefereinheit,
-  menge: BigDecimal,
-  preis: Option[BigDecimal],
-  anzahl: Int,
-  //modification flags
-  erstelldat: DateTime,
-  ersteller: PersonId,
-  modifidat: DateTime,
-  modifikator: PersonId
-) extends BaseEntity[BestellpositionId]
-
-case class BestellpositionModify(
-  bestellungId: BestellungId,
-  produktId: Option[ProduktId],
-  produktBeschrieb: String,
-  preisEinheit: Option[BigDecimal],
-  einheit: Liefereinheit,
-  menge: BigDecimal,
-  preis: Option[BigDecimal],
-  anzahl: Int
-) extends JSONSerializable
-
-case class ProduzentenabrechnungReport(
-  produzentId: ProduzentId,
-  produzentKurzzeichen: String,
-  produzent: ProduzentDetailReport,
-  bestellungenDetails: Seq[BestellungDetail],
-  preisTotal: BigDecimal,
-  steuerSatz: Option[BigDecimal],
-  steuer: BigDecimal,
-  totalSteuer: BigDecimal,
-  //Zusatzinformationen
-  projekt: ProjektReport
-) extends JSONSerializable
-
-case class KorbId(id: Long) extends BaseId
-
-case class Korb(
-  id: KorbId,
-  lieferungId: LieferungId,
-  aboId: AboId,
-  status: KorbStatus,
-  guthabenVorLieferung: Int,
-  auslieferungId: Option[AuslieferungId],
-  sort: Option[Int],
-  //modification flags
-  erstelldat: DateTime,
-  ersteller: PersonId,
-  modifidat: DateTime,
-  modifikator: PersonId
-) extends BaseEntity[KorbId]
-
-case class KorbLieferung(
-  id: KorbId,
-  lieferungId: LieferungId,
-  aboId: AboId,
-  status: KorbStatus,
-  guthabenVorLieferung: Int,
-  auslieferungId: Option[AuslieferungId],
-  sort: Option[Int],
-  lieferung: Lieferung,
-  //modification flags
-  erstelldat: DateTime,
-  ersteller: PersonId,
-  modifidat: DateTime,
-  modifikator: PersonId
-) extends JSONSerializable
-
-case class KorbDetail(
-  id: KorbId,
-  lieferungId: LieferungId,
-  abo: Abo,
-  status: KorbStatus,
-  guthabenVorLieferung: Int,
-  auslieferungId: Option[AuslieferungId],
-  sort: Option[Int],
-  kunde: Kunde,
-  abotyp: Abotyp,
-  //modification flags
-  erstelldat: DateTime,
-  ersteller: PersonId,
-  modifidat: DateTime,
-  modifikator: PersonId
-) extends JSONSerializable
-
-case class KorbReport(
-  id: KorbId,
-  lieferungId: LieferungId,
-  abo: Abo,
-  status: KorbStatus,
-  guthabenVorLieferung: Int,
-  auslieferungId: Option[AuslieferungId],
-  sort: Option[Int],
-  kunde: KundeReport,
-  abotyp: Abotyp,
-  //modification flags
-  erstelldat: DateTime,
-  ersteller: PersonId,
-  modifidat: DateTime,
-  modifikator: PersonId
-) extends JSONSerializable
-
-case class KorbModify(
-  id: KorbId
-) extends JSONSerializable
-
-case class KorbCreate(
-  LieferungId: LieferungId,
-  aboId: AboId,
-  status: KorbStatus,
-  guthabenVorLieferung: Int
 ) extends JSONSerializable
